Split the client into multiple classes inherited
