+++ lib/highline_wrapper.rb
@@ -9,7 +9,7 @@ Dir[files].each do |file|
 end
 
 class HighlineWrapper
-  ASK_DEFAULTS = {
+  OPEN_ENDED_DEFAULTS = {
     secret: false,
     required: false
   }.freeze
@@ -38,8 +38,8 @@ class HighlineWrapper
   # Notes:
   #  If required == true, the question will repeat until the user answers the question
   def ask(prompt, options = {})
-    options = ASK_DEFAULTS.merge(options)
-    client.ask(prompt, options)
+    options = OPEN_ENDED_DEFAULTS.merge(options)
+    HighlineWrapper::OpenEndedQuestion.ask(prompt, options)
   end
 
   # Returns: yes for true, no for false (boolean)
@@ -54,7 +54,7 @@ class HighlineWrapper
   #  If required == true, then the default value will be ignored
   def ask_yes_no(prompt, options = {})
     options = YES_NO_DEFAULTS.merge(options)
-    client.ask_yes_no(prompt, options)
+    HighlineWrapper::YesNoQuestion.ask(prompt, options)
   end
 
   # Returns: the selection in a hash (hash)
@@ -74,7 +74,7 @@ class HighlineWrapper
   #   If default == nil and required == false, and the user skips the question, the answer will be nil
   def ask_multiple_choice(prompt, choices, options = {})
     options = MULTIPLE_CHOICE_DEFAULTS.merge(options)
-    client.ask_multiple_choice(prompt, choices, options)
+    HighlineWrapper::MultipleChoiceQuestion.ask(prompt, choices, options)
   end
 
   # Returns: the selections chosen as an array of hashes (array)
@@ -94,10 +94,6 @@ class HighlineWrapper
   #   If defaults == [] and required == false, then the method will return an empty array
   def ask_checkbox(prompt, choices, options = {})
     options = CHECKBOX_DEFAULTS.merge(options)
-    client.ask_checkbox(prompt, choices, options)
-  end
-
-  private def client
-    @client ||= HighlineWrapper::Client.new
+    HighlineWrapper::CheckboxQuestion.ask(prompt, choices, options)
   end
 end
